# okrTrackingSystem/api-docs.yaml
openapi: 3.1.0
info:
  title: OKR Tracking System API
  description: |
    ## OKR Tracking System — Garant Bank

    REST API for managing **Objectives & Key Results (OKR)**, organizational structure, user management, and multi-source performance evaluations.

    ### Key Features
    - **OKR Management** — Create and track Objectives with weighted Key Results
    - **Multi-Source Evaluations** — Director (star rating), HR (letter grade), and Business Block assessments
    - **Score Calculation** — Automatic weighted scoring: 60% OKR + 20% Director + 20% HR
    - **Organizational Hierarchy** — Divisions → Departments → Objectives → Key Results
    - **Role-Based Access Control** — 6 roles with fine-grained permissions
    - **Excel Export** — Export all OKR data to Excel spreadsheets

    ### Authentication
    All endpoints (except login and export) require a **JWT Bearer token**.
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click the **Authorize** button above and enter the token

    ### Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Full system access, user management |
    | `DIRECTOR` | Division oversight, star-based evaluations |
    | `DEPARTMENT_LEADER` | Department OKR management |
    | `HR` | Letter-grade evaluations (A/B/C/D) |
    | `BUSINESS_BLOCK` | Business block star-based evaluations |
    | `EMPLOYEE` | View own profile and assigned departments |
  version: 1.0.0
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by ID
      description: Returns user details by ID. ADMIN and DIRECTOR can view any user.
        Other users can only view their own profile.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "403":
          description: Can only view own profile
        "404":
          description: User not found
    put:
      tags:
      - Users
      summary: Update user
      description: |-
        Update user details.

        **ADMIN** can update all fields including role, department assignments, and account status.

        **Regular users** can only update their own profile fields: fullName, email, jobTitle, phoneNumber, bio, password.
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: Invalid request
        "403":
          description: Cannot update admin-only fields
        "404":
          description: User not found
    delete:
      tags:
      - Users
      summary: Delete user
      description: Permanently delete a user account. Users cannot delete their own
        account. **Requires ADMIN role.**
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: User ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: User deleted
        "400":
          description: Cannot delete your own account
        "403":
          description: Only ADMIN can delete users
        "404":
          description: User not found
...
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register new user
      description: Create a new user account. **Requires ADMIN role.** The user will
        be assigned the specified role and optionally linked to a department.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "400":
          description: "Invalid request (e.g., duplicate username)"
        "403":
          description: Only ADMIN can register users
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: User login
      description: Authenticate with username and password to receive a JWT token.
        The token should be included in the `Authorization` header as `Bearer <token>`
        for all subsequent requests.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Login successful — returns JWT token and user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
      security: []
...
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user
      description: Returns the profile information of the currently authenticated
        user based on the JWT token.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDTO"
        "401":
          description: Not authenticated
...
components:
  schemas:
    UpdateUserRequest:
      type: object
      description: Request to update user details. All fields are optional — only
        provided fields are updated. Fields marked 'ADMIN only' are ignored for non-admin
        users.
      properties:
        fullName:
          type: string
          description: Full display name
          example: John Doe Updated
        email:
          type: string
          description: Email address
          example: john.updated@garantbank.uz
        jobTitle:
          type: string
          description: Job title
          example: Lead Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 987 65 43
        bio:
          type: string
          description: Bio/description
          example: Team lead with 5 years experience
        role:
          type: string
          description: User role (ADMIN only)
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: DEPARTMENT_LEADER
        assignedDepartmentIds:
          type: array
          description: Department IDs to assign (ADMIN only)
          example:
          - dept-001
          - dept-002
          items:
            type: string
        password:
          type: string
          description: "New password (optional, for password change)"
          example: newSecurePass456
        isActive:
          type: boolean
          description: Account active status (ADMIN only)
          example: true
        canEditAssignedDepartments:
          type: boolean
          description: Whether user can edit assigned departments (ADMIN only)
          example: true
        readOnly:
          type: boolean
          description: Whether user is in read-only mode (ADMIN only)
          example: false
    DepartmentSummaryDTO:
      type: object
      description: Lightweight department reference
      properties:
        id:
          type: string
          description: Department ID
          example: dept-001
        name:
          type: string
          description: Department name
          example: Software Development
    UserDTO:
      type: object
      description: User information (password excluded)
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: Unique username
          example: john.doe
        email:
          type: string
          description: Email address
          example: john.doe@garantbank.uz
        fullName:
          type: string
          description: Full display name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        profilePhotoUrl:
          type: string
          description: URL to user's profile photo
          example: /uploads/photo-abc123.jpg
        jobTitle:
          type: string
          description: Job title
          example: Senior Developer
        phoneNumber:
          type: string
          description: Phone number
          example: +998 90 123 45 67
        bio:
          type: string
          description: User bio/description
          example: Experienced developer focused on backend systems
        lastLogin:
          type: string
          format: date-time
          description: Timestamp of last login
          example: 2025-01-15T10:30:00
        assignedDepartments:
          type: array
          description: Departments the user is assigned to
          items:
            $ref: "#/components/schemas/DepartmentSummaryDTO"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2025-01-01T09:00:00
        active:
          type: boolean
        isActive:
          type: boolean
          description: Whether the user account is active
          example: true
        canEditAssignedDepartments:
          type: boolean
          description: Whether the employee can edit their assigned departments
          example: false
        readOnly:
          type: boolean
          description: Whether the user is in read-only mode (cannot make any edits)
          example: false
...
    RegisterRequest:
      type: object
      description: User registration request (admin only)
      properties:
        username:
          type: string
          description: Unique username
          example: john.doe
        email:
          type: string
          description: Email address
          example: john.doe@garantbank.uz
        password:
          type: string
          description: Password (min 6 characters)
          example: securePass123
        fullName:
          type: string
          description: User's full display name
          example: John Doe
        role:
          type: string
          description: User role
          enum:
          - EMPLOYEE
          - DEPARTMENT_LEADER
          - HR
          - DIRECTOR
          - BUSINESS_BLOCK
          - ADMIN
          example: EMPLOYEE
        departmentId:
          type: string
          description: Optional department ID to assign
          example: dept-001
      required:
      - email
      - fullName
      - password
      - role
      - username
    LoginRequest:
      type: object
      description: Login credentials
      properties:
        username:
          type: string
          description: Username
          example: admin
        password:
          type: string
          description: Password
          example: admin123
      required:
      - password
      - username
    LoginResponse:
      type: object
      description: Login response containing JWT token and user details
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiJ9...
        type:
          type: string
          default: Bearer
          description: Token type
          example: Bearer
        user:
          $ref: "#/components/schemas/UserDTO"
          description: Authenticated user details
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter the JWT token obtained from `POST /api/auth/login`
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT